chore: Add function to convert feature collection to GeoDataFrame
This commit adds a new function `fc2gdf` to the `core.py` file. This function takes a feature collection from Earth Engine and converts it to a GeoDataFrame using the `GeoDataFrame.from_features` method from the `geopandas` library. This function is useful for converting Earth Engine feature collections to a format that can be easily used and analyzed in Python.