feat: Add function to convert image collection to GeoDataFrame
This commit introduces a new function `image_collection_to_dataframe` in the `core.py` file. This function takes an image collection from Earth Engine and converts it to a GeoDataFrame using the `fc2gdf` function from the `gee2gpd` module. This functionality allows for easy conversion and analysis of Earth Engine image collections in Python.